Add an axis to an r3js plot

This is used as part of the plot3js() function but can be called separately to add an axis, generally in combination after other lower level functions like plot3js.new() and plot3js.window().

axis3js(
  data3js,
  side,
  at = NULL,
  labels = NULL,
  cornerside = "f",
  labeloffset = 0.1,
  ...
)

Arguments

data3js

The data3js object

side

The axis side, either "x", "y" or "z"

at

Where to draw labels

labels

Vector of labels to use

cornerside

See material3js()

labeloffset

Amount of offset of axis labels from the edge of the plot

...

Other arguments to pass to material3js()

Value

Returns an updated data3js object

See also