Set material properties of an r3js object

Arguments refer to different material properties for an object, many of which refer directly to properties as described in the 'threejs' documentation

material3js(
  mat = "phong",
  col = "black",
  fill = "black",
  opacity = NULL,
  xpd = TRUE,
  lwd = 1,
  dashSize = NULL,
  gapSize = NULL,
  interactive = NULL,
  label = NULL,
  toggle = NULL,
  depthWrite = NULL,
  depthTest = NULL,
  polygonOffset = NULL,
  polygonOffsetFactor = NULL,
  polygonOffsetUnits = NULL,
  shininess = 30,
  faces = NULL,
  corners = NULL,
  rotation = NULL,
  normalise = NULL,
  poffset = NULL,
  clippingPlanes = NULL,
  frontSide = TRUE,
  backSide = TRUE,
  renderOrder = NULL,
  ...
)

Arguments

mat

Material to use for the object, one of "basic", "lambert", "phong" or "line", see e.g. MeshBasicMaterial

col

Color

fill

Fill color

opacity

Opacity

xpd

Should parts of the object outside the plot limits be shown

lwd

Line width

dashSize

Dash size for dashed lines

gapSize

Gap size for dashed lines

interactive

Is the object interactive

label

The label for the object

toggle

Toggle button associated with the object

depthWrite

See depthWrite

depthTest

See depthTest

polygonOffset

See polygonOffset

polygonOffsetFactor

See polygonOffsetFactor

polygonOffsetUnits

See polygonOffsetUnits

shininess

Shininess of object surface

faces

For dynamically hidden objects, the face with which it is associated, see details.

corners

For dynamically hidden objects, the corners with which it is associated, see details.

rotation

In place rotation of the object geometry (most relevant for points)

normalise

Should coordinates be normalised to be with respect to axis ranges or placed according to the plotting box which has unit coordinates.

poffset

Positional offset, the offset is relative to the plotting area size rather than axis limits

clippingPlanes

Clipping planes to apply to the object

frontSide

Logical indicating whether the front side of a mesh should be rendered

backSide

Logical indicating whether the back side of a mesh should be rendered

renderOrder

See renderOrder

...

Additional arguments (not used)

Value

Returns a list of material properties