Add a side grid to the plot

This is used by the function grid3js to add a grid on the faces of the plot region, marking for example axis points.

sidegrid3js(
  data3js,
  axis,
  side,
  at = NULL,
  col = "grey80",
  dynamic = FALSE,
  geometry = FALSE,
  ...
)

Arguments

data3js

The data3js object

axis

Axis for which to draw lines, one of "x", "y" or "z"

side

The side of the grid, one of "x", "y" or "z"

at

Where to draw grid lines, defaults to the current axis tick marks

col

Color of the grid lines

dynamic

Should edges of the box closest to the viewer hide themselves automatically

geometry

Should the lines be rendered as a physical geometry in the scene (see lines3js())

...

Other arguments to pass to material3js()